Revert type files move.
