+++ src/worker/thread-worker.ts
@@ -1,7 +1,7 @@
 import { isMainThread, parentPort } from 'worker_threads'
 import type { MessageValue } from '../utility-types'
+import type { WorkerOptions } from '../worker-options'
 import { AbstractWorker } from './abstract-worker'
-import type { WorkerOptions } from './worker-options'
 
 /**
  * An example worker that will be always alive, you just need to **extend** this class if you want a static pool.
